Some changes to ensure compatibility with gcc 10
 On branch feature/gcc9
 Changes to be committed:
	modified:   Addons/HoughTransform/include/DHoughTransform.hpp
	modified:   CMakeLists.txt
	modified:   Core/include/DBaseImage.h
	modified:   Morpho/include/private/DMorphoDistance.hpp

+++ Addons/HoughTransform/include/DHoughTransform.hpp
@@ -123,7 +123,7 @@ namespace smil
 
     size_t wOut = xyResol * wIn;
     size_t hOut = xyResol * hIn;
-    size_t dOut = rhoResol * rhoMax;
+    off_t  dOut = rhoResol * rhoMax;
 
     ImageFreezer freeze(imOut);
     imOut.setSize(wOut, hOut, dOut);
@@ -133,7 +133,7 @@ namespace smil
     typename Image<T1>::lineType lIn;
     typename Image<T2>::volType slicesOut = imOut.getSlices();
 
-    size_t rho;
+    off_t rho;
     double coef = 1 / (xyResol * rhoResol);
 
 #ifdef USE_OPEN_MP
